pyproject.toml
[project]
name = "quickapp"
version = "1.0.0"
description = "AI-Powered Web App Generator using FastAPI + Jinja2 + SQLite"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic-ai>=0.0.49",
    "fastapi>=0.115.0",
    "jinja2>=3.1.0",
    "mistralai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
quickapp = "quickapp:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
